Recycler list works plus minor ui changes

+++ app/src/main/java/uk/ac/abertay/cmp309/assesment/Shop.kt
@@ -4,7 +4,7 @@ data class Shop(
     val Name: String ?=null,
     val Image: String ?=null,
     val Distance: String ?=null,
-    val Rating: Long ?= null
+    val Rating: Double ?= null
 
 
 )
